+++ src/View/View/AppView.php
@@ -23,13 +23,12 @@
 namespace MeCms\View\View;
 
 use Cake\Routing\Router;
-use MeCms\View\View\BaseView;
-use MeTools\Core\Plugin;
+use MeCms\View\View;
 
 /**
  * Application view class for all views, except the admin views
  */
-class AppView extends BaseView
+class AppView extends View
 {
     /**
      * Internal property to set the userbar elements
@@ -42,7 +41,7 @@ class AppView extends BaseView
      * Internal method to set some blocks
      * @return void
      * @uses $userbar
-     * @uses MeCms\View\View\BaseView::_getTitleForLayout()
+     * @uses MeCms\View\View::_getTitleForLayout()
      * @uses MeTools\View\Helper\HtmlHelper::meta()
      * @uses MeTools\View\Helper\LibraryHelper::analytics()
      * @uses MeTools\View\Helper\LibraryHelper::shareaholic()
@@ -92,7 +91,7 @@ class AppView extends BaseView
      * Initialization hook method
      * @return void
      * @see http://api.cakephp.org/3.3/class-Cake.View.View.html#_initialize
-     * @uses MeCms\View\View\BaseView::initialize()
+     * @uses MeCms\View\View::initialize()
      */
     public function initialize()
     {
@@ -113,33 +112,12 @@ class AppView extends BaseView
      * @param string|null $layout Layout name
      * @return mixed Rendered output, or false on error
      * @see http://api.cakephp.org/3.3/class-Cake.View.View.html#_renderLayout
-     * @uses MeCms\View\View\BaseView::renderLayout()
-     * @uses MeTools\Core\Plugin::path()
+     * @uses MeCms\View\View::renderLayout()
      * @uses _setBlocks()
      * @uses userbar()
      */
     public function renderLayout($content, $layout = null)
     {
-        $path = 'Template' . DS . 'Layout' . DS;
-
-        if ($this->layoutPath()) {
-            $path .= $this->layoutPath() . DS;
-        }
-
-        $path .= $layout . '.ctp';
-
-        $this->plugin = MECMS;
-
-        //Uses the APP layout, if exists
-        if (is_readable(APP . $path)) {
-            $this->plugin = false;
-        }
-
-        //Sets the theme and uses the theme layout, if exists
-        if ($this->theme() && is_readable(Plugin::path($this->theme()) . 'src' . DS . $path)) {
-            $this->plugin = $this->theme();
-        }
-
         $this->_setBlocks();
 
         //Assign the userbar