+++ tests/test_app/TestApp/Plugin/TestPlugin/src/View/Cell/PluginExampleWidgetsCell.php
@@ -3,7 +3,7 @@ namespace TestPlugin\View\Cell;
 
 use Cake\View\Cell;
 
-class PluginExampleCell extends Cell
+class PluginExampleWidgetsCell extends Cell
 {
     public function display()
     {